+++ src/main/kotlin/cloud/glitchdev/rfu/feature/fishing/FrogcoinBlessingFeature.kt
@@ -1,9 +1,10 @@
 package cloud.glitchdev.rfu.feature.fishing
 
-import cloud.glitchdev.rfu.events.managers.ChatEvents.registerChatEvent
+import cloud.glitchdev.rfu.events.managers.ChatEvents.registerGameEvent
 import cloud.glitchdev.rfu.events.managers.TickEvents.registerTickEvent
 import cloud.glitchdev.rfu.feature.Feature
 import cloud.glitchdev.rfu.feature.RFUFeature
+import cloud.glitchdev.rfu.utils.dsl.toExactRegex
 import kotlin.time.Duration.Companion.minutes
 
 @RFUFeature
@@ -12,11 +13,11 @@ object FrogcoinBlessingFeature : Feature {
     val activeBlessings: Map<String, Long>
         get() = _activeBlessings
 
-    private val blessingRegex = Regex("""^WISE! You've been granted \+(.+?) for (\d+)m while on the Lotus Atoll!$""")
+    private val blessingRegex = """WISE! You've been granted \+(.+?) for (\d+)m while on the Lotus Atoll!""".toExactRegex()
 
     override fun onInitialize() {
-        registerChatEvent(filter = blessingRegex) { _, match ->
-            val matchResult = match ?: return@registerChatEvent
+        registerGameEvent(filter = blessingRegex) { _, _, match ->
+            val matchResult = match ?: return@registerGameEvent
             val buffInfo = matchResult.groupValues[1]
             val durationMinutes = matchResult.groupValues[2].toLongOrNull() ?: 30L